Rename all GameSrc and MacSrc files into *.cpp
Technically, this is C++ project now, right?

+++ src/GameSrc/Headers/treasure.h
@@ -18,7 +18,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 // You got nothing. Zero.
-#define NOTHING_TRIPLE 0xFFFFFFFF
+#define NOTHING_TRIPLE static_cast<int>(0xFFFFFFFF)
 #define NUM_TREASURE_TYPES 15
 #define NUM_TREASURE_SLOTS 7
 #define NUM_TREASURE_ENTRIES 2
